Added Session, Navigation, Activity helpers
  Added login method to session helper
  Moved up all the test helpers to new class
  Created navigation class helpers to handle navigation
