+++ src/test/java/appManager/SessionHelper.java
@@ -1,0 +1,22 @@
+package appManager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class SessionHelper {
+    private WebDriver driver;
+
+    public SessionHelper(WebDriver driver) {
+        this.driver = driver;
+    }
+
+    public void login(String username, String passowrd) {
+        driver.findElement(By.id("username")).click();
+        driver.findElement(By.id("username")).clear();
+        driver.findElement(By.id("username")).sendKeys(username);
+        driver.findElement(By.id("password")).click();
+        driver.findElement(By.id("password")).clear();
+        driver.findElement(By.id("password")).sendKeys(passowrd);
+        driver.findElement(By.xpath("//button[@type='submit']")).click();
+    }
+}
